Added Cooldowns

+++ src/main/java/me/hektortm/woSSystems/utils/dataclasses/Cooldown.java
@@ -2,11 +2,11 @@ package me.hektortm.woSSystems.utils.dataclasses;
 
 public class Cooldown {
     private final String id;
-    private final int duration;
+    private final long duration;
     private final String start_interaction;
     private final String end_interaction;
 
-    public Cooldown(String id, int duration, String start_interaction, String end_interaction) {
+    public Cooldown(String id, long duration, String start_interaction, String end_interaction) {
         this.id = id;
         this.duration = duration;
         this.start_interaction = start_interaction;
@@ -17,7 +17,7 @@ public class Cooldown {
         return id;
     }
 
-    public int getDuration() {
+    public long getDuration() {
         return duration;
     }
